added to routes for supplement details and packs. Issue #125
